completed watchlist, authorization, default image, styling

+++ commerce/auctions/views.py
@@ -9,44 +9,52 @@ from django.contrib.auth.decorators import login_required
 from django.db import connection
 
 
+defaultimage = "/static/auctions/images/default.png" 
+
 def dictfetchall(cursor):
 
     columns = [col[0] for col in cursor.description]
     return [dict(zip(columns, row)) for row in cursor.fetchall()]
 
 def index(request, viewtype='active'):
+	print(defaultimage)
 	title = ""
+	userid = "0"
+	userauth = False
 	if request.user.is_authenticated:
-		cursor = connection.cursor()
-		# convert to int first to prevent any possible userid spoofing sql injection
-		userid = str(int(request.user.id))
-	#	sSQL = """select l.id 'id', l.title 'title', l.image_link, l.description, l.listing_date, l.startbid, max(b.bid_amount) 'bid_amount', count(b.listing_id) 'bid_count' 
-	#			from auctions_listing l 
-	#			left join auctions_listingbid b on l.id = b.listing_id 
-	#			group by l.id,b.listing_id;"""
-		if viewtype=='active':
-			sSQL = f"""select l.id 'id', l.title 'title', l.image_link, l.description, l.listing_date, l.startbid, max(b.bid_amount) 'bid_amount', count(b.listing_id) 'bid_count', case when w.id is not null then 'true' else 'false' end 'watchlist' 
-						from auctions_listing l 
-						left join auctions_listingbid b on l.id = b.listing_id
-						left join auctions_watchlist w on w.listing_id_id  = l.id and w.user_id_id = {userid}
-						group by l.id,b.listing_id, w.id;"""
-			title = "Active Listings"
-			cursor.execute(sSQL)		
-		else: 
+		userauth = True
+	cursor = connection.cursor()
+	
+	if request.user.is_authenticated:
+		userid = str(request.user.id)
+	
+	if viewtype=='add':
+		return newlisting(request, 'add')
+	if viewtype=='active':
+		title = "Active Listings"
+		cursor.execute("""select l.id 'id', l.title 'title', l.image_link, l.description, l.listing_date, l.startbid, max(b.bid_amount) 'bid_amount', count(b.listing_id) 'bid_count', case when w.id is not null then 'true' else 'false' end 'watchlist' 
+					from auctions_listing l 
+					left join auctions_listingbid b on l.id = b.listing_id
+					left join auctions_watchlist w on w.listing_id_id  = l.id and w.user_id_id = %s
+					where l.active = true
+					group by l.id,b.listing_id, w.id;""", [userid])		
+	else: 
+		if request.user.is_authenticated:
 			title = "Watchlist"
-
-			sSQL = (f"""select l.id 'id', l.title 'title', l.image_link, l.description, l.listing_date, l.startbid, max(b.bid_amount) 'bid_amount', count(b.listing_id) 'bid_count' 
+			cursor.execute("""select l.id 'id', l.title 'title', l.image_link, l.description, l.listing_date, l.startbid, max(b.bid_amount) 'bid_amount', count(b.listing_id) 'bid_count' 
 					from auctions_listing l 
 					left join auctions_listingbid b on l.id = b.listing_id 
 					inner join auctions_Watchlist w on w.listing_id_id = l.id 
-					where w.user_id_id = {userid} 
-					group by l.id,b.listing_id;""")			
-			cursor.execute(sSQL)		
-			#cursor.execute("select l.id 'id', l.title 'title', l.image_link, l.description, l.listing_date, l.startbid, max(b.bid_amount) 'bid_amount', count(b.listing_id) 'bid_count' from auctions_listing l left join auctions_listingbid b on l.id = b.listing_id group by l.id,b.listing_id;")		
-		listings = dictfetchall(cursor)  #Listing.objects.raw("select l.id 'id', l.title 'title', l.image_link, l.startbid, b.bid_amount 'bid_amount', b.bid_count 'bid_count' from auctions_listing l left join (select listing_id, max(bid_amount) 'bid_amount', count(listing_id) 'bid_count' from auctions_listingbid group by listing_id) b on l.id = b.listing_id;")
-		return render(request, "auctions/index.html", {"listings": listings, "title": title, "viewtype":viewtype})
-	else:
-		return render(request, "auctions/login.html", {"message": "Login required."})
+					where w.user_id_id = %s 
+					group by l.id,b.listing_id;""", [userid])		
+		else:
+			return render(request, "auctions/login.html", {"message": "Login required."})
+	listings = dictfetchall(cursor)  
+	
+	return render(request, "auctions/index.html", {"listings": listings, "userauth":userauth, "title": title, "viewtype":viewtype})
+	#else:
+		#return render(request, "auctions/login.html", {"message": "Login required."})
+	#	return HttpResponseRedirect(reverse("index"))
 
 def login_view(request):
     if request.method == "POST":
@@ -69,9 +77,10 @@ def login_view(request):
 
 
 def logout_view(request):
-    logout(request)
-    return HttpResponseRedirect(reverse("index"))
-
+	print('logout')
+	logout(request)
+	#return HttpResponseRedirect(reverse("index", args=('logout',)))
+	return render(request, "auctions/login.html")
 
 def register(request):
     if request.method == "POST":
@@ -99,17 +108,20 @@ def register(request):
     else:
         return render(request, "auctions/register.html")
         
+@login_required(login_url="/login")
 def newlisting(request, formtype, listingid=None):
-	print(formtype)
 	if request.method == "POST":
 		if formtype=="add":
+			if request.POST["imageurl"].strip() == "":
+				imageurl = "/static/auctions/images/default.png" 
+			else:
+				imageurl = 	equest.POST["imageurl"].strip()
 			categoryid = int(request.POST["category"])
 			category = Category.objects.get(id=categoryid)
 			listing = Listing.objects.create(user_id=request.user, title=request.POST["title"], description=request.POST["description"], 
-			category_id=category, image_link=request.POST["imageurl"], listing_date=datetime.now(), active=True, startbid=float(request.POST["startbid"]))
+			category_id=category, image_link=imageurl, listing_date=datetime.now(), active=True, startbid=float(request.POST["startbid"]))
 			return HttpResponseRedirect(reverse("index"))
 		if formtype=="edit":
-			print(listingid)
 			listing = Listing.objects.get(id=int(request.POST["listingid"]))
 			categoryid = int(request.POST["category"])
 			listing.category = Category.objects.get(id=categoryid)
@@ -123,14 +135,14 @@ def newlisting(request, formtype, listingid=None):
 			return HttpResponseRedirect(reverse("index"))
 				
 	elif formtype=="edit":
-		categories = Category.objects.all()
+		categories = Category.objects.all().order_by('category_name').values()
 		formlabel = "Edit"
 		listing = Listing.objects.get(id=listingid)
 		print(listing.description)
 		return render(request, "auctions/newlisting.html", { "categories": categories, "formlabel": formlabel, "scontent": listing.description, 
 		"title": listing.title, "selectedcategory": listing.category_id, "image": listing.image_link, "formtype": formtype, "listingid":listingid, "startbid":listing.startbid})	
 	else:	
-		categories = Category.objects.all()
+		categories = Category.objects.all().order_by('category_name').values()
 		formlabel = "Create"
 		return render(request, "auctions/newlisting.html", { "categories": categories, "formlabel": formlabel, "formtype": formtype})	
 
@@ -138,6 +150,15 @@ def viewlisting(request, listingid):
 	listing = Listing.objects.get(id=listingid)
 	listingbids = ListingBid.objects.filter(listing = listing)
 	bidcount = listingbids.count()
+	bwatchlist = False
+	userauth = False
+	if request.user.is_authenticated:
+		userauth = True
+		#listinguser = Listing.user_id
+		watchlist = Watchlist.objects.filter(listing_id=listing, user_id=request.user)
+		for item in watchlist:
+			if item.listing_id == listing:
+				bwatchlist = True
 	if(bidcount > 0):
 		lastbid = listingbids.last().bid_amount
 		lastbiduser = listingbids.last().user_id
@@ -145,13 +166,9 @@ def viewlisting(request, listingid):
 		lastbid = 0
 		lastbiduser = ""
 	comments = Comment.objects.filter(listing_id = listing)
-	listinguser = Listing.user_id
-	#return render(request, "auctions/viewlisting.html", { "categories": categories, "formlabel": formlabel, "scontent": listing.description, 
-	#"title": listing.title, "selectedcategory": listing.category_id, "image": listing.image_link, "formtype": formtype, "listingid":listingid, "startbid":listing.startbid,
-	#"comments":comments, "listinguser":listinguser})	
-	return render(request, "auctions/viewlisting.html", {"listing": listing, "listingbids": lastbid, "comments":comments, "bidscount":bidcount, "lastbiduser":lastbiduser})
+	return render(request, "auctions/viewlisting.html", {"listing": listing, "userauth":userauth, "watchlist": bwatchlist, "listingbids": lastbid, "comments":comments, "bidscount":bidcount, "lastbiduser":lastbiduser})
 	
-@login_required
+@login_required(login_url="/login")
 def bid(request, listingid):
 	if request.method == "POST":	
 		listing = Listing.objects.get(id=listingid)
@@ -160,7 +177,7 @@ def bid(request, listingid):
 	else:
 		return viewlisting(request, listingid) 
 		
-@login_required
+@login_required(login_url="/login")
 def comment(request, listingid):
 	if request.method == "POST":	
 		listing = Listing.objects.get(id=listingid)
@@ -169,15 +186,26 @@ def comment(request, listingid):
 	else:
 		return viewlisting(request, listingid) 
 		
-@login_required
+@login_required(login_url="/login")
 def watchlist(request, listingid, remove=0):
+		print(request)
 		listing = Listing.objects.get(id=listingid)
-		if remove > 0:
+		if remove > 0 and remove < 4:
 			Watchlist.objects.filter(user_id=request.user, listing_id = listing).delete()
-		else:	
+		elif remove == 4 or remove==0:	
 			Watchlist.objects.create(user_id=request.user, listing_id = listing)
+		if remove >= 3:
+			return viewlisting(request, listingid)
 		if remove == 1:
 			return index(request, 'watchlist') 	
 		else:
 			return index(request, 'active')	
+
+def close(request, listingid):
+	if request.method == "POST":
+		listing = Listing.objects.get(id=listingid)
+		listing.active = False
+		listing.save()
+	return index(request)	
+
